requirements.txt
streamlit
pycryptodome
tinyec
